requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy
scikit-learn
joblib==1.3.2
mysql-connector-python 
python-dotenv